Refactor: input.js에서 도메인 관련 로직 분리
